Support changing a team name in ways that avoid changing the slug
It's possible to change a team name in ways that avoid changing the
slug. For example, changing capitalization, adding accent marks, or
inserting symbols that are stripped from slugs (e.g. the `#` symbol).
If a team name is changed in one of these ways, we can preserve the
existing GitHub Team and only update its name. (On the other hand, if
a team name is changed in ways that also change the slug, the action
will always create a new team with a new slug. To avoid that, users
can rename the GitHub Team before updating the teams document with the
new team name.)