Validate teams configuration and parse it into a Map<string, TeamData>
This ensures the config we loaded has the expected structure, which is
an object mapping team names to team metadata. The metadata must contain
a `members` array of objects which must include a `github` property
whose value is a GitHub username. The metadata can also optionally
include a `description` string and/or `team_sync_ignored` boolean. Any
other fields may be present, but they'll be ignored by this action.